runtime/mail_api/pyproject.toml
[project]
name = "psense-mail-api"
version = "0.1.0"
description = "PSense Mail — Configuration-driven Python FastAPI backend"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "PSense.ai" }]

dependencies = [
    # --- Web framework ---
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "python-multipart>=0.0.18",

    # --- MongoDB ---
    "motor>=3.6,<4",
    "beanie>=1.27,<2",

    # --- Config ---
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "pyyaml>=6,<7",

    # --- Auth (KeyCloak JWT) ---
    "python-jose[cryptography]>=3.3,<4",
    "httpx>=0.28,<1",

    # --- Mail transport ---
    "aiosmtplib>=3,<4",

    # --- File I/O ---
    "aiofiles>=24,<25",

    # --- Utilities ---
    "structlog>=24,<25",
    "python-ulid>=3,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6,<7",
    "mongomock-motor>=0.0.34",
    "ruff>=0.8,<1",
]
s3 = ["aiobotocore>=2,<3"]
gmail = ["google-api-python-client>=2,<3", "google-auth>=2,<3", "google-auth-oauthlib>=1,<2"]
azure = ["azure-storage-blob[aio]>=12,<13"]
gcs = ["gcloud-aio-storage>=9,<10"]

[tool.hatch.build.targets.wheel]
packages = ["app", "config"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
